requirements.txt
django

djangorestframework

django-cors-headers

djangorestframework-simplejwt

psycopg2

gunicorn

whitenoise

django-environ

django-debug-toolbar

pytest

pytest-django

factory_boy

faker